+++ compiler/src/main/java/info/teksol/mc/profile/options/MlogFormatOptions.java
@@ -10,4 +10,7 @@ public enum MlogFormatOptions {
     NO_ARGUMENT_PADDING,
     SIGNATURE,
     SYMBOLIC_LABELS,
+    PROCESSOR_ID,
+    PROGRAM_NAME,
+    PROGRAM_VERSION,
 }